Hide buffer instead of wiping it
When the buffer is exited, hide it automatically. When :Oldfiles is
called again, use the same buffer but repopulate it.

This commit also includes a fix for CTRL-^ not working after selecting a
file from the Oldfiles buffer. After changing files, the alternate file
register (@#) was pointing to the Oldfiles buffer, which was now wiped,
so pressing CTRL-^ gave the message "No alternate file". Instead, after
selecting a file from the buffer, force @# to be what it was before
opening the Oldfiles buffer.