Only fire autocmds on non-empty filenames
The ? character in the autocmd pattern matches a single character, so
the pattern `?*` is similar to `.+` in regex; essentially, it matches
on one or more characters.

This prevents the autocmds from firing on empty buffers. Both
oldfiles#add() and oldfiles#remove() are no-ops on empty buffers anyway,
so by ignoring these patterns in the autocmds we can defer the
autoloading until the functions are *actually* needed.