Expand filenames before filtering
Vim uses a tilde for the user's home directory in the oldfiles list
which returns false when passed to the filereadable function. This was
causing all file paths beneath the user's home directory to be filtered
out. To fix this, expand all paths before passing to the filter function
so that we are always working with absolute paths.